Implement movement range charting
Determine how far someone can move from a given
starting point in a given number of steps.

Start at some point an then recursively work
outwards until the number of steps run out.

(This is possibly not the fastest implementaion.
Let's get back to this if/when speed starts becomming an issue)