Add  type TerrainPalette to module 'EdgeDetection
Add container type that match edge combinations
with a tile. Uses a Dict wrapped in helper functions
that convert EdgeSet |> toString.

This hack lets us use an enumerated type as the key of
a dict dispite not counting as a 'comparable' in elm.
Sealing the type and accessing it with functions creates
adequate type safety.

Next up: Craft terrain from a Set of Coords.