AGBIC2018 - Introduce new flavours: Chilli and Chocolate
Complement Sugar with two additional flavours: Chilli and Chocolate
This is intendet to make the board a bit more stable,
as it lessens the number of things that can mix with each other.

Flour, Water and Bun(s) now come in various flavours.
Flour and Water of different flavours do not mix.
To produce a bun of some flavour, the player must combine either
Flour and Water of the same flavour, or
one with the intended flavour and the other without flavour.