AGBIC2018 - Allow a Spawner to spawn different things
Spawner now contains a list of different things that it could spawn.
One of them is chosen at random when spawning.

Unify spawn points in level so that all spawners spawn all things.
This prevents large stack of Flour and Warer from piling up
at the right and left side of the level.

Also refactor function spawnSingelThingRnd so that it returns
a ```Maybe ( Coords, Thingy )``` instead of a ```Grid Thingy```.
This makes it easier to write tests and is generally less confusing.