Prevent swap, unless is causes a chain reaction
Modify SweetBuns.attemptMove to only preform a swap if the
result is not a stable grid.
Also recactor the whole function for improved readability.

Simplify signaturs of:
  - SweetBuns.applyGravity
  - SweetBuns.isStable

Bouns:
Renamed SweetBuns.move to SweetBuns.moveMixning to clarify that
it will attempt to mix things if the destination cell
is already occupied.